move pending txs to a context
