+++ packages/ui/src/contexts/ModalsContext.tsx
@@ -1,14 +1,13 @@
-import { useState, useContext, createContext, useCallback, useMemo } from 'react'
+import { useState, useContext, createContext, useCallback } from 'react'
 import ChangeMultisig from '../components/modals/ChangeMultisig'
 import EditNames from '../components/modals/EditNames'
 import Send, { DEFAULT_EASY_SETUP_SELECTION, EasyTransferTitle } from '../components/modals/Send'
-import { usePendingTx } from '../hooks/usePendingTx'
 import { SignClientTypes } from '@walletconnect/types'
 import WCSessionProposal from '../components/modals/WalletConnectSessionProposal'
 import ProposalSigningModal, { SigningModalProps } from '../components/modals/ProposalSigning'
 import WalletConnectSigning from '../components/modals/WalletConnectSigning'
-import { useMultiProxy } from './MultiProxyContext'
 import HiddenAccountInfo from '../components/modals/HiddenAccountInfo'
+import { usePendingTx } from './PendingTxContext'
 
 interface ModalsContextProps {
   setIsEditModalOpen: (isOpen: boolean) => void
@@ -48,13 +47,8 @@ const ModalsContextProvider = ({ children }: React.PropsWithChildren) => {
   const [wCSessionProposal, setWCSessionProposal] = useState<
     OpenWCModalParams['sessionProposal'] | undefined
   >()
-  const { selectedMultiProxy } = useMultiProxy()
-  const multisigAddresses = useMemo(
-    () => selectedMultiProxy?.multisigs.map(({ address }) => address) || [],
-    [selectedMultiProxy?.multisigs]
-  )
-  const { refresh } = usePendingTx({ multisigAddresses, withPplChain: false })
-  const { refresh: refreshPpl } = usePendingTx({ multisigAddresses, withPplChain: true })
+
+  const { refresh, refreshPpl } = usePendingTx()
   const onCloseEditModal = useCallback(() => setIsEditModalOpen(false), [setIsEditModalOpen])
   const onCloseChangeMultiModal = useCallback(
     () => setIsChangeMultiModalOpen(false),
